# docs/swagger.yaml
definitions:
  addy.Alias:
    properties:
      active:
        type: boolean
      aliasable_id: {}
      aliasable_type: {}
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      domain:
        type: string
      email:
        type: string
      emails_blocked:
        type: integer
      emails_forwarded:
        type: integer
      emails_replied:
        type: integer
      emails_sent:
        type: integer
      extension: {}
      from_name: {}
      id:
        type: string
      last_blocked: {}
      last_forwarded:
        type: string
      last_replied: {}
      last_sent: {}
      local_part:
        type: string
      recipients:
        items: {}
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  addy.AliasFormat:
    enum:
    - random_characters
    - uuid
    - random_words
    - custom
    type: string
    x-enum-varnames:
    - AliasFormatRandomCharacters
    - AliasFormatUUID
    - AliasFormatRandomWords
    - AliasFormatCustom
  addy.CreateAliasReq:
    properties:
      authorization:
        type: string
      description:
        type: string
      domain:
        type: string
      format:
        $ref: '#/definitions/addy.AliasFormat'
      local_part:
        type: string
      recipient_ids:
        items:
          type: string
        type: array
      xrequestedWith:
        type: string
    type: object
  addy.CreateAliasResp:
    properties:
      data:
        $ref: '#/definitions/addy.Alias'
    type: object
  addy.ErrorResp:
    properties:
      error:
        type: string
    type: object
  enum.ProviderEnum:
    enum:
    - addy
    - sl
    type: string
    x-enum-varnames:
    - ProviderEnumAddy
    - ProviderEnumSimpleLogin
...
  v1.Alias:
    properties:
      alias:
        type: string
      callLogId:
        type: string
      id:
        type: integer
      mockProvider:
        $ref: '#/definitions/enum.ProviderEnum'
      targetEmail:
        type: string
      tokenId:
        type: string
    type: object
...
  v1.ListAliasResp:
    properties:
      list:
        items:
          $ref: '#/definitions/v1.Alias'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  v1.Response:
    properties:
      code:
        description: Code 指定了业务错误码.
        type: string
      data:
        description: Data 包含了
      message:
        description: Message 包含了可以直接对外展示的错误信息.
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/aliases:
    get:
      consumes:
      - application/json
      description: Get a list of aliases with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        maximum: 200
        minimum: 1
        name: pageSize
        type: integer
      - collectionFormat: csv
        description: Order by fields
        in: query
        items:
          type: string
        name: orderBy
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/v1.ListAliasResp'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/v1.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.Response'
      summary: List aliases
      tags:
      - aliases
    post:
      consumes:
      - application/json
      description: Create a new email alias using the Addy.io API
      parameters:
      - description: API Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Alias creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/addy.CreateAliasReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/addy.CreateAliasResp'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/addy.ErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/addy.ErrorResp'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/addy.ErrorResp'
      summary: Create an alias for Addy.io
      tags:
      - bridges
...
swagger: "2.0"